+++ shigg/__init__.py
@@ -13,6 +13,7 @@ from .elements import Draggable
 from .elements import MoveAndResizeThumbs
 from .elements import Label
 from .elements import LeftRightSelector
+from .elements import ButtonToggle
 
 # export utils
 from .utils import transform_mouse_to_normalized_subsurface_coords
@@ -25,3 +26,4 @@ from .default_drawing import draw_move_and_resize_thumbs
 from .default_drawing import draw_label
 from .default_drawing import draw_left_right_selector
 from .default_drawing import draw_vertical_slider
+from .default_drawing import draw_button_toggle