+++ shigg/default_drawing/__init__.py
@@ -6,4 +6,5 @@ from .pygame import (
     draw_move_and_resize_thumbs,
     draw_label,
     draw_left_right_selector,
+    draw_button_toggle,
 )