+++ shigg/elements/__init__.py
@@ -5,6 +5,7 @@ from ._element_event import ElementEvent
 # elements
 from .button import Button
 from .slider import Slider
+from .button_toggle import ButtonToggle
 from .vertical_slider import VerticalSlider
 from .draggable import Draggable
 from .move_and_resize_thumbs import MoveAndResizeThumbs