+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## New in 0.2.3
+### New Elements
+- added a button based toggle element (ButtonToggle)
+- added additional test scenarios
+
 ## New in 0.2.2
 ### Bug Fixes
 - fixed vertical slider click box being wider than the thumb; now correct size
@@ -6,7 +11,7 @@
 
 ## New in 0.2.1
 ### New Elements
-- Added a VerticalSlider element
+- added a VerticalSlider element
 - added horizontal list selector element (LeftRightSelector)
 - added a static label element